# apps/relayer/openapi-spec.json
{"openapi":"3.0.0","paths":{"/api/relayer/relay/evm-evm":{"post":{"operationId":"relayEvmToEvm","summary":"Relay a message from EVM chains","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelayerEvmRequest"}}}},"responses":{"201":{"description":""}},"tags":["relayer"]}},"/api/relayer/relay/xrpl-evm":{"post":{"operationId":"relayXrplToEvm","summary":"Relay a message from XRPL chains","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelayerXrplRequest"}}}},"responses":{"201":{"description":""}},"tags":["relayer"]}},"/api/relayer/relay/evm-xrpl":{"post":{"operationId":"relayEvmToXrpl","summary":"Relay a message from EVM chains","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelayerEvmRequest"}}}},"responses":{"201":{"description":""}},"tags":["relayer"]}}},"info":{"title":"relayer","description":"Relayer API","version":"0.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ApiException":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"RelayerEvmRequest":{"type":"object","properties":{"sourceChain":{"type":"string"},"sourceAddress":{"type":"string"},"txHash":{"type":"string"},"txEvent":{"type":"string"},"payloadHash":{"type":"string"},"destinationChain":{"type":"string"},"destinationAddress":{"type":"string"}},"required":["sourceChain","sourceAddress","txHash","txEvent","payloadHash","destinationChain","destinationAddress"]},"XrplUserRequest":{"type":"object","properties":{"tx_id":{"type":"string"},"sourceAddress":{"type":"string"},"destinationChain":{"type":"string"},"destinationAddress":{"type":"string"},"amount":{"type":"string","description":"The amount to be relayed (in drops)"},"payloadHash":{"type":"string"}},"required":["tx_id","sourceAddress","destinationChain","destinationAddress","amount","payloadHash"]},"RelayerXrplRequest":{"type":"object","properties":{"user_request":{"allOf":[{"$ref":"#/components/schemas/XrplUserRequest"}]},"sourceChain":{"type":"string"}},"required":["user_request","sourceChain"]}}}}